requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
django-environ==0.9.0
environ==1.0
idna==3.4
msal==1.20.0
Office365-REST-Python-Client==2.3.16
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7
requests==2.28.1
urllib3==1.26.13
